+++ frontend/app/page.tsx
@@ -1,38 +1,26 @@
-"use client";
+import type { Metadata } from "next";
+import ClientPage from "./client_page";
 
-import ScrollableContentContainer from "./components/scrollable_content_container";
-import SplashPage from "./components/splash_page";
-import TopBarSplashPage from "./components/top_bar_splash_page";
-import { useSearchParams } from "next/navigation";
-import { useState } from "react";
-import RateLimitExceededModal from "./components/rate_limit_exceeded_modal";
-import { useRouter } from "next/navigation";
-import Link from "next/link";
+export const metadata: Metadata = {
+  title: "Active Statistics",
+  description:
+    "Active Statistics is an open source visualiser of Strava data. Connect your Strava account and visualise your data in new and exciting ways.",
+  openGraph: {
+    title: "Active Statistics",
+    description:
+      "Active Statistics is an open source visualiser of Strava data. Connect your Strava account and visualise your data in new and exciting ways.",
+    url: "https://active-statistics.com/",
+    images: [
+      {
+        url: "https://active-statistics.com/active_statistics_2.png",
+        width: 1200,
+        height: 630,
+        alt: 'A pixelated image with the words "Active Statistics" and then a stick drawing of a person running.',
+      },
+    ],
+  },
+};
 
 export default function Index() {
-  const searchParams = useSearchParams();
-  const [modalState, setModalState] = useState(
-    searchParams.get("rate_limit_exceeded") == "true",
-  );
-
-  const router = useRouter();
-
-  return (
-    <>
-      <TopBarSplashPage />
-      <ScrollableContentContainer>
-        <SplashPage />
-      </ScrollableContentContainer>
-      <RateLimitExceededModal
-        handleExit={() => {
-          setModalState(false);
-          router.push("/");
-        }}
-        modalState={modalState}
-      ></RateLimitExceededModal>
-      <div className="text-center mb-2">
-        Made with ❤️ in <Link className="hyperlink" href="https://en.wikipedia.org/wiki/Brisbane">Brisbane</Link> by <Link className="hyperlink" href="https://johnscolaro.xyz">John</Link>
-      </div>
-    </>
-  );
+  return <ClientPage />;
 }
